Adapted panel to be used as a drawer, starter preparation for recreating element form
