+++ lib/data/services/bricks_generator_service.dart
@@ -2,7 +2,7 @@ import 'dart:developer';
 import 'dart:io';
 
 import 'package:flutter_arhitect/common/models/arhitecture_elements/base_arhitecture_element.dart';
-import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:mason/mason.dart';
 
 final bricksServiceProvider = Provider<BricksGeneratorService>(
