+++ lib/data/services/database_service.dart
@@ -1,6 +1,6 @@
 import 'package:flutter_arhitect/common/models/arhitecture_elements/base_arhitecture_element.dart';
 import 'package:flutter_arhitect/common/models/arhitecture_elements/element_parts/arhitecture_layer.dart';
-import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:hooks_riverpod/hooks_riverpod.dart';
 
 final databaseServiceProvider = Provider<DatabaseService>((ref) {
   return DatabaseServiceImpl();
