+++ lib/domain/arhitecture_element_pairs_provider.dart
@@ -1,7 +1,7 @@
 import 'package:flutter_arhitect/common/models/element_pair.dart';
 import 'package:flutter_arhitect/domain/all_arhitecture_elements_notifier.dart';
 import 'package:flutter_arhitect/extensions/global_key_extension.dart';
-import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:hooks_riverpod/hooks_riverpod.dart';
 
 final arhitectureElementPairsProvider = Provider<List<ElementPair>>((ref) {
   final allArhitectureElements = ref.watch(allArhitectureElementsNotifier);
