+++ lib/domain/currently_selected_arhitecture_element_state_provider.dart
@@ -2,13 +2,13 @@ import 'package:dartz/dartz.dart';
 import 'package:flutter/widgets.dart';
 import 'package:flutter_arhitect/common/models/arhitecture_elements/base_arhitecture_element.dart';
 import 'package:flutter_arhitect/domain/all_arhitecture_elements_notifier.dart';
-import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:hooks_riverpod/hooks_riverpod.dart';
 
 final currentlySelectedArhitectureElementStateProvider =
     StateProvider<BaseArhitectureElement?>((ref) => null);
 
 final currentMethodsAndParametersStateProvider =
-    StateProvider<List<Tuple2<Widget, List<Widget>>>>((ref) => []);
+    StateProvider.autoDispose<List<Tuple2<Widget, List<Widget>>>>((ref) => []);
 
 final arhitectureElementProvider =
     Provider.family<BaseArhitectureElement, String>((ref, id) {
