+++ lib/domain/files_generation_notifier.dart
@@ -3,7 +3,7 @@ import 'dart:developer';
 import 'package:flutter_arhitect/data/repositories/bricks_repository.dart';
 import 'package:flutter_arhitect/domain/all_arhitecture_elements_notifier.dart';
 import 'package:flutter_arhitect/domain/common/request_notifier/request_notifier.dart';
-import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:hooks_riverpod/hooks_riverpod.dart';
 
 final filesGenerationNotifier =
     StateNotifierProvider<FilesGenerationNotifier, void>((ref) {
