+++ lib/domain/mouse_position_state_provider.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:hooks_riverpod/hooks_riverpod.dart';
 
 final mousePositionStateProvider = StateProvider<Offset>((ref) {
   return Offset.zero;
