+++ lib/domain/selected_widgets_notifier.dart
@@ -3,7 +3,7 @@ import 'dart:developer';
 import 'package:flutter_arhitect/common/models/arhitecture_elements/base_arhitecture_element.dart';
 import 'package:flutter_arhitect/domain/all_arhitecture_elements_notifier.dart';
 import 'package:flutter_arhitect/extensions/global_key_extension.dart';
-import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:hooks_riverpod/hooks_riverpod.dart';
 
 final selectedWidgetsNotifier =
     StateNotifierProvider<SelectedWidgetsNotifier, SelectedElements>((ref) {
