+++ lib/main.dart
@@ -5,7 +5,7 @@ import 'package:flutter_arhitect/domain/selected_widgets_notifier.dart';
 import 'package:flutter_arhitect/domain/user_connecting_positions_provider.dart';
 import 'package:flutter_arhitect/presentation/painters/user_connecting_line_painter.dart';
 import 'package:flutter_arhitect/scale_update_details_state_provider.dart';
-import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:hooks_riverpod/hooks_riverpod.dart';
 
 const projectName = 'starter';
 const featureName = 'login';
