+++ lib/presentation/pannel/pannel_body.dart
@@ -1,14 +1,13 @@
 import 'dart:developer';
 
 import 'package:flutter/material.dart';
-import 'package:flutter_arhitect/common/models/arhitecture_elements/base_arhitecture_element.dart';
 import 'package:flutter_arhitect/domain/all_arhitecture_elements_notifier.dart';
-import 'package:flutter_arhitect/domain/currently_selected_arhitecutre_element_state_provider.dart';
+import 'package:flutter_arhitect/domain/currently_selected_arhitecture_element_state_provider.dart';
 import 'package:flutter_arhitect/forms/architecture_element_form.dart';
 import 'package:flutter_arhitect/presentation/widgets/custom_text_field.dart';
-import 'package:flutter_arhitect/presentation/widgets/methods_and_parameters_form_widget.dart';
+import 'package:flutter_arhitect/presentation/widgets/methods_and_parameters_form/methods_and_parameters_form_widget.dart';
 import 'package:flutter_form_builder/flutter_form_builder.dart';
-import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:hooks_riverpod/hooks_riverpod.dart';
 
 final formKeyProvider = Provider<GlobalKey<FormBuilderState>>((ref) {
   return GlobalKey<FormBuilderState>();
@@ -55,9 +54,11 @@ class _Body extends ConsumerWidget {
   @override
   Widget build(BuildContext context, WidgetRef ref) {
     final initialValue = ref.read(
-          currentlySelectedArhitectureElementStateProvider,
-        ) ??
-        BaseArhitectureElement.empty();
+      currentlySelectedArhitectureElementStateProvider,
+    );
+    if (initialValue == null) {
+      return const SizedBox();
+    }
     final arhitectureElement =
         ref.watch(arhitectureElementProvider(initialValue.id));
 
@@ -76,6 +77,7 @@ class _Body extends ConsumerWidget {
               CustomTextField.normal(
                 name: ArchitectureElementForm.nameKey,
                 text: 'Name',
+                initialValue: initialValue.name,
                 autoValidateMode: AutovalidateMode.onUserInteraction,
                 isRequired: true,
                 // isRequiredValidatorErrorText: 'Name is required',
@@ -86,6 +88,7 @@ class _Body extends ConsumerWidget {
               CustomTextField.multilineDescription(
                 name: ArchitectureElementForm.descriptionKey,
                 text: 'Description',
+                initialValue: initialValue.description,
                 textInputType: TextInputType.multiline,
                 isRequired: false,
               ),
@@ -93,6 +96,7 @@ class _Body extends ConsumerWidget {
               TextButton(
                 child: const Text('Save'),
                 onPressed: () {
+                  log('formKey.currentState!.value: ${formKey.currentState!.value}');
                   if (formKey.currentState?.saveAndValidate() == true) {
                     ref
                         .read(allArhitectureElementsNotifier.notifier)
