+++ pubspec.yaml
@@ -37,16 +37,12 @@ dependencies:
   flutter_form_builder: ^7.8.0
   flutter_hooks: ^0.18.6
   flutter_reorderable_list: ^1.2.0
-  flutter_riverpod: ^2.2.0
   form_builder: ^1.0.1
   form_builder_validators: ^8.5.0
   freezed_annotation: ^2.2.0
   hooks_riverpod: ^2.3.1
   mason: ^0.1.0-dev.47
   mason_cli: ^0.1.0-dev.49
-  riverpod: ^2.2.0
-  riverpod_annotation: ^2.0.0
-  riverpod_generator: ^2.0.0
   uuid: ^3.0.7
 
 dev_dependencies:
